Add custom exceptions, implement global exception handler, refactor s…

+++ src/main/java/com/example/application/order/OrderService.java
@@ -2,6 +2,7 @@ package com.example.application.order;
 
 import com.example.domain.order.IOrderRepo;
 import com.example.domain.order.Order;
+import com.example.exception.customException.ResourceNotFound;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -25,16 +26,28 @@ public class OrderService implements IOrderService{
 
     @Override
     public Order getOrderById(UUID id) {
-        return orderRepo.getOrderById(id);
+        Order order = orderRepo.getOrderById(id);
+        if (order == null) {
+            throw new ResourceNotFound("Order not found with id: " + id);
+        }
+        return order;
     }
 
     @Override
     public Order updateOrder(UUID id, Order order) {
+        Order existingOrder = orderRepo.getOrderById(id);
+        if (existingOrder == null) {
+            throw new ResourceNotFound("Order not found with id: " + id);
+        }
         return orderRepo.updateOrder(id, order);
     }
 
     @Override
     public void deleteOrder(UUID id) {
+        Order order = orderRepo.getOrderById(id);
+        if (order == null) {
+            throw new ResourceNotFound("Order not found with id: " + id);
+        }
         orderRepo.deleteOrder(id);
     }
 }
